Update: app.js
I fixed the select button issue by separating it into a separate Div called "selectDiv", and putting the image and recipe name in another div called "subDiv", and appending both to the "recipeDiv". So the pushPage happens when "subDiv" is clicked; the select action happens when user clicks "selectDiv".

I also constructed a rough Recipe Detail window to use in place of the iFrame, just to demonstrate what it might contain. Obviously, it needs to be formatted. But it would include a (larger) image, title, rating, ingredients, etc. Then all of our data would be consistent and we wouldn't be linking to all these different websites (and generating a slew of errors in the console each time).